Update README.md
If you are on this team, please give your best effort. To succeed we need to work effectively as a team. If theres any question or needing assistace dont hessitate just ask. We're a team, all for one and all for all. United theres no stoping us. Im going to pay for everyones(and a guest i.e. significant other, friend, sibbling etc.. trip to celebrate our sucess to  (yet to be determined). A weekend of celebration and to our future lives getting better in many ways. Be confident, believe, we can achieve. I will soon reveal 3 locations and everyone can vote, that way the team remains a team in every decision this project makes! Lets do this!